Create object detection using the model's output
defined the function detect_objects() to perform object detection using the model's output. Also, it's important to note that the code above is just an example, you may need to adjust the pre-processing steps based on the format of the image or video and the requirements of your model.

Also, you can use libraries like OpenCV to work with images and videos instead of PIL, and you can use libraries like cv2.imread() to read an image or cv2.VideoCapture() to read a video.
assuming that   the data is pre-trained object detection model and have defined the function detect_objects() to perform object detection using the model's output. 

Also, you can use libraries like OpenCV to work with images and videos instead of PIL, and you can use libraries like cv2.imread() to read an image or cv2.VideoCapture() to read a video.